spring boot + shardingsphere + mybatis plus 分库分表集成完成

+++ cv-cloud-db/cv-shardingsphere-sample/src/main/java/com/clearvision/shardingspheresample/common/config/DatabaseShardingConfig.java
@@ -8,10 +8,11 @@ import org.springframework.stereotype.Component;
 import java.util.Collection;
 
 @Component
-public class DatabaseShardingConfig implements PreciseShardingAlgorithm<Integer>  {
+public class DatabaseShardingConfig implements PreciseShardingAlgorithm<Long>  {
     @Override
-    public String doSharding(Collection<String> availableTargetNames, PreciseShardingValue<Integer> preciseShardingValue) {
-        Integer value = preciseShardingValue.getValue();
+    public String doSharding(Collection<String> availableTargetNames, PreciseShardingValue<Long> preciseShardingValue) {
+
+        Long value = preciseShardingValue.getValue();
         String shardingValue = String.valueOf(value);
         for (String targetName : availableTargetNames) {
             if(targetName.endsWith(shardingValue)) {
